requirements.txt
fastapi
uvicorn[standard]
python-multipart
pillow
torch
timm
numpy